🛠 rename: Move actions, functions files into a folder that corresponds to each function
